refactor(refactor): move Reactor to core namespace
