+++ include/core/Reactor.hpp
@@ -12,7 +12,7 @@
 #include "shared/defines.hpp"
 #include "www/VirtualServer.hpp"
 
-namespace www {
+namespace core {
 
 /**
  * @class Reactor
@@ -25,11 +25,11 @@ class Reactor {
 
   void init() throw(std::exception);
   int getEpollFd() const;
-  const std::vector<VirtualServer>& getVirtualServers() const;
-  std::vector<VirtualServer>& getVirtualServers();
+  const std::vector<www::VirtualServer>& getVirtualServers() const;
+  std::vector<www::VirtualServer>& getVirtualServers();
 
   void addVirtualServer(config::t_server& serverConfig) throw(std::exception);
-  bool removeVirtualServer(std::vector<VirtualServer>::iterator it);
+  bool removeVirtualServer(std::vector<www::VirtualServer>::iterator it);
   bool addVirtualServers(config::t_config_data& configData);
 
   void registerHandler(int fd, runtime::RequestHandler* handler, uint32_t events = EPOLLIN) throw(std::runtime_error);
@@ -41,8 +41,8 @@ class Reactor {
   Reactor& operator=(const Reactor& other);
 
   int m_epoll_master_fd;
-  std::vector<VirtualServer> m_virtual_servers;
+  std::vector<www::VirtualServer> m_virtual_servers;
   std::map<int, runtime::RequestHandler*> m_event_handlers;
 };
 
-} /* namespace www */
+}  // namespace core
