+++ src/main.cpp
@@ -1,8 +1,8 @@
 #include <iostream>
 
 #include "config/Parser.hpp"
+#include "core/Reactor.hpp"
 #include "logger/LoggerManager.hpp"
-#include "www/Reactor.hpp"
 
 int main(int argc, char** const argv) {
   try {
@@ -20,7 +20,7 @@ int main(int argc, char** const argv) {
     } else {
       configfile = "config/configfile_example";
     }
-    www::Reactor server;
+    core::Reactor server;
     config::ConfigFileParser configFileParser;
     if (configFileParser.loadConfigFile(configfile) == 1) return 1;
